+++ NEWS.md
@@ -21,3 +21,9 @@
 * Version name: "Schrödinger's cat jumps out of the box"
 * Added correction factor on LPSMAP covariance matrix.
 * episim() can now simulate from negative binomial distribution.
+
+### Version 1.0.6 ### (**2022-07-28**)
+
+* Version name: "Welcome to the Metaverse"
+* Added midwindow option to the perfcheck routine. It reports EpiEstim Rt 
+  estimates at the middle of the sliding window.